requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
alabaster==1.0.0
attrs==25.4.0
babel==2.17.0
blinker==1.9.0
cachelib==0.13.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
docutils==0.22.4
Flask==3.1.2
Flask-Session==0.8.0
frozenlist==1.8.0
gunicorn==23.0.0
idna==3.11
imagesize==1.4.1
itsdangerous==2.2.0
Jinja2==3.1.6
libsql-client==0.3.1
MarkupSafe==3.0.3
msgspec==0.20.0
multidict==6.7.0
packaging==25.0
pdfkit==1.0.0
propcache==0.4.1
Pygments==2.19.2
python-dotenv==1.2.1
requests==2.32.5
roman-numerals==4.1.0
snowballstemmer==3.0.1
Sphinx==9.0.4
sphinx-press-theme==0.8.0
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
typing_extensions==4.15.0
urllib3==2.6.2
Werkzeug==3.1.4
yarl==1.22.0
